backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
websockets==12.0
gtts==2.5.0
pydantic==2.5.2
aiofiles==23.2.1
python-multipart==0.0.6
